Add files via upload
The file ENSP_simulation 1, ENSP_simulation 2 and ENSP_simulation3 are the core program of the simulation. Firstly, please ensure eNSP is downloaded and installed following the instruction. The official website for download is https://support.huawei.com/enterprise/en/network-management/ensp-pid-9017384  . On 8 July, we  accessed the website but found that no software is available. Since the installation package is 413 MB, which is more than the limit size of github.  An upload of the installation is not possible. We thereby provide another address in  website (sorry not in English), in which the software can be found and available for download: https://www.ttrar.com/html/152975.html. To download it, an account for Baidu is necessary. After download, the software can be installed according to the "installation instruction".